Add components for chess analysis dashboard
- Implement competitor analysis component to display current Elo ratings and predicted win probabilities.
- Create game length analysis component to visualize game duration statistics and results.
- Develop landing page component for initial user interaction and information.
- Add opening performance component to analyze results across different chess openings.
- Introduce opponent strength component to categorize opponents and visualize win rates.
- Create performance overview component to summarize user statistics and recent games.
- Implement win probability component to model and visualize win probabilities based on historical data.
- Add rating trend component to track user rating progression over time.
- Create results over time component to aggregate and visualize game outcomes by week.
- Implement sidebar component for navigation and data management, including game fetching and filtering options.
